refactor(publisher): one owner for failed-job policy; split the 2270 suite
GH#2270 PR-2, review round 3 (#2293). No behaviour change except the last item.

The policy predicates were spread across two modules while each read the same failed job and
each could answer for the others — the shape helpers drift into. `async-lift-retry-disposition.ts`
is now their canonical owner, with the precedence documented once at the top: proven-ineffective →
held for chain proof → retry action → occupancy and bulk clearability → the read projection.
`async-lift-publisher-utils.ts` keeps the STRUCTURAL helpers (is this job failed, what does it
persist, how is it rebuilt) and says where the rest went. `hasBroadcastEvidence` moved with them:
reading the fields is structural, but treating them as EVIDENCE is a judgement.

The public surface shrinks to the read model. `classifyRetryAction`, `describeRetryProjection`,
`deriveLiftJobRetryProjection` and the action-string union leave the barrel: a consumer reads a
job's `retryState` off the publisher and the counts off `retryDetailed`, so nothing outside this
package re-derives either. `LiftJobRetryProjection`, `LiftJobRetryWaitingReason`,
`AsyncLiftRetryOutcome` and `LiftJobPendingChainProofError` stay public. No CLI import changed.

The 1021-line spec file split along its seams, every file registered in the unit config:
- `async-lift-auto-retry-2270` — the automatic lane (17 rows);
- `async-lift-retry-disposition-2270` — evidence, the hold, the classifier, the projection (11);
- `async-lift-admission-clear-2270` — what a held job does to admission and cleanup (8 + 1 new).
The clock, ids, store and the four failure fixtures moved to `_helpers/async-lift-2270-harness.ts`,
so `advance()` and `freshStore()` are the only places a row moves shared state. Row accounting:
627 → 628 in the publisher lane, the one added row being the new admission case.

That new row is the dangerous intersection of the two admission rules: an evidence-bearing job
whose budget is SPENT. The fresh-mandate path re-arms an exhausted budget and re-runs the same
job — which for a transaction nobody has accounted for republishes it — while refusing it the
wrong way (letting the subject go) mints a replacement and publishes twice. The hold wins over
both, and the row pins that the budget is not re-armed and no replacement appears. The exhausted
state is written directly onto the persisted job because the lane can never spend a held job's
budget; that IS the hold.

Mutants: neutering the writer's hold guard fails 4 rows including both admission cases (so the
new row is a real discriminator, not a restatement of its neighbour).
