+++ packages/publisher/vitest.unit.config.ts
@@ -23,6 +23,8 @@ export default defineConfig({
       'test/async-lift-writejob-atomicity.test.ts',
       'test/async-lift-publish-options.test.ts',
       'test/async-lift-auto-retry-2270.test.ts',
+      'test/async-lift-retry-disposition-2270.test.ts',
+      'test/async-lift-admission-clear-2270.test.ts',
       'test/lift-job-failure.test.ts',
       'test/async-promote-queue.test.ts',
       'test/async-promote-writejob-atomicity.test.ts',
